create attendance input component
